upd: rebase & continue splitting components
- layouts: header & footer
- components: basic buttton
- pages: login page
